chore: refactor startup
